+++ src/ManageCourses.Api/Services/Data/IDataService.cs
@@ -6,7 +6,6 @@ namespace GovUk.Education.ManageCourses.Api.Data
 {
     public interface IDataService
     {
-        void ProcessUcasPayload(UcasPayload payload);
         IEnumerable<UserOrganisation> GetOrganisationsForUser(string email);
         UserOrganisation GetOrganisationForUser(string email, string instCode);
         Course GetCourse(string email, string instCode, string ucasCode);
